move the ids file to the root package
