+++ resotocore/resotocore/task/task_description.py
@@ -21,7 +21,7 @@ from resotocore.message_bus import Event, Action, ActionDone, Message, ActionErr
 from resotocore.model.typed_model import to_json, from_js, to_js
 from resotocore.types import Json
 from resotocore.util import first, interleave, empty, exist, identity, utc, utc_str
-from resotocore.model.ids import SubscriberId
+from resotocore.ids import SubscriberId
 
 
 log = logging.getLogger(__name__)
